pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "haymaker"
authors = [
        {name = "t1user", email = "enterdragon@duck.com"},
]
description = 'Framework for trading with Interactive Brokers through ib_insync'
readme = "README.md"
requires-python = ">= 3.8"
keywords = ["ib_insync", "tws", "ibgateway", "asyncio",  "interactive brokers", "algo trading"]
classifiers = [
            "Development Status :: 3 - Alpha",
            "Intended Audience :: Developers",
            "Topic :: Office/Business :: Financial :: Investment",
            "Licence :: OSI Approved :: BSD Licence",
            "Programming Language :: Python :: 3",

]
dependencies = [
    "arctic @ git+https://github.com/man-group/arctic.git",
    "ib_insync",
    "pandas",
    "numba",
    "pyyaml",
]
dynamic = ["version"]

[project.optional-dependencies]
notebook = [
         "jupyter",
         "matplotlib",
         "pyfolio"
         ]
dev = [
    "types-pyyaml",
    "pandas-stubs",
    "isort",
    "mypy",
    "autopep8",
    "yapf",
    "black",
    "flake8",
    "pytest",
    "pytest-asyncio"
    ]
all = ["haymaker[notebook,dev]"]


[tool.setuptools]
packages = ["haymaker"]

[project.urls]
Repository = "https://github.com/t1user/haymaker"

[project.scripts]
dataloader = "haymaker.dataloader.dataloader:start"

[tool.setuptools_scm]
version_file = "haymaker/_version.py"

[tool.isort]
profile = "black"

[tool.mypy]
mypy_path = "$MYPY_CONFIG_FILE_DIR/stubs"
# check_untyped_defs = true

# remove this after fixing backtester
[[tool.mypy.overrides]]
module = "backtester"
ignore_errors = true

[tool.pytest.ini_options]
addopts = ["-p", "no:warnings"]

